ark pytorch validation
