+++ examples/tutorial/torch_tutorial.py
@@ -3,123 +3,25 @@
 
 import ark
 import torch
-import torch.optim as optim
-
-from ark.module import ARKComponent, ARKFunction, ARKLayer
-from ark import Tensor, Parameter
 
 
-
-# Define a custom ARK function for a linear layer
-class ARKLinear(ARKFunction):
-    @staticmethod
-    def forward(ark_input, ark_weight):
-        return ark.matmul(ark_input, ark_weight, transpose_other=True)
-
-    @staticmethod
-    def setup_context(ctx, inputs, output):
-        input, weight = inputs
-        ctx.save_for_backward(input, weight)
-
-    @staticmethod
-    def backward(ctx, ark_grad_output, ark_input, ark_weight):
-        grad_input = grad_weight = None
-        if ctx.needs_input_grad[0]:
-            grad_input = ark.matmul(
-                ark_grad_output, ark_weight, transpose_other=False
-            )
-        if ctx.needs_input_grad[1]:
-            grad_weight = ark.matmul(
-                ark_input, ark_grad_output, transpose_input=True
-            )
-        return grad_input, grad_weight
-
-ark_linear = ARKLinear()
-
-# Define a PyTorch model
-class SimpleModel(torch.nn.Module):
-    def __init__(self):
-        super().__init__()
-        self.layers = torch.nn.Sequential(
-            torch.nn.Linear(256, 256, bias=False),  # Layer 0
-            torch.nn.Linear(256, 256, bias=False),  # Layer 1
-            torch.nn.Linear(256, 256, bias=False),  # Layer 2
-            torch.nn.Linear(256, 256, bias=False),  # Layer 3
-            torch.nn.Linear(256, 256, bias=False),  # Layer 4
-            
-        )
-
-    def forward(self, x):
-        return self.layers(x)
-
-def replace_layers_with_ark(model, layer_indicies):
-    weight = model.layers[0].weight.clone().detach().to("cuda:0")
-    ark_weight = Parameter.from_tensor(Tensor.from_torch(weight.requires_grad_(True)))
-
-    weight_1 = model.layers[1].weight.clone().detach().to("cuda:0")
-    ark_weight_1 = Parameter.from_tensor(Tensor.from_torch(weight_1.requires_grad_(True)))
-
-    model.layers[0] = ARKComponent([
-        ARKLayer(ark_linear, ark_weight),
-        ARKLayer(ark_linear, ark_weight_1)
-    ])
-    del model.layers[1]
-    
-    return model
+class ArkAddModule(ark.RuntimeModule):
+    def build_forward(self, x: ark.Tensor, y: ark.Tensor) -> ark.Tensor:
+        return ark.add(x, y)
 
 
+# ARK module for addition
+module = ArkAddModule()
 
-# Move the PyTorch model to GPU
-pytorch_model = SimpleModel()
-ark_model = SimpleModel()
-ark_model.load_state_dict(pytorch_model.state_dict())
-ark_model = replace_layers_with_ark(ark_model, [0])
+# Define two torch arrays
+x = torch.ones(64) * 2
+y = torch.ones(64) * 3
 
-# Let's print the layers of our PyTorch model and the hybrid model
-print("PyTorch model:\n", pytorch_model)
-print("\nARK model:\n", ark_model)
+# Run the ARK module
+z = module(x, y)
 
+w = module(x, z)
 
-# Move the hybrid to GPU
-pytorch_model.to("cuda:0")
-ark_model.to("cuda:0")
-
-# Now lets run the model on some random input
-input_tensor = torch.randn(128, 256).to("cuda:0").requires_grad_(True)
-input2 = input_tensor.clone().detach().requires_grad_(True)
-
-# Compare the results of both models
-
-# Define an arbitrary target
-target = torch.randn(128, 256).to("cuda:0") 
-
-loss_fn = torch.nn.MSELoss()
-optim_torch = optim.SGD(pytorch_model.parameters(), lr=0.01)
-optim_ark = optim.SGD(ark_model.parameters(), lr=0.01)
-num_iters = 1
-for iter in range(num_iters):
-    print(f"Iteration {iter+1}/{num_iters}")
-
-    optim_torch.zero_grad()
-    optim_ark.zero_grad()
-
-    pytorch_output = pytorch_model(input_tensor)
-    ark_output = ark_model(input2)
-
-    print("PyTorch output:", pytorch_output)
-    print("ARK output:", ark_output)
-
-    #Compute loss
-    torch_loss = loss_fn(pytorch_output, target)
-    ark_loss = loss_fn(ark_output, target)
-
-    print(f"PyTorch loss: {torch_loss.item()}")
-    print(f"ARK loss: {ark_loss.item()}")
-
-    # Perform a backward pass
-    torch_loss.backward()
-    ark_loss.backward()
-
-    optim_torch.step()
-    optim_ark.step()
-
+# Print the result
+print(z)  # 5
+print(w)  # 7
